pyproject.toml
[project]
name = "nod-detector"
version = "0.1.0"
description = "A system for detecting nodding behavior in videos"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Kagan Kucukaytekin", email = "kkaytekin@gmail.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.20.0,<2.0.0",
    "opencv-python>=4.5.1.48",
    "mediapipe>=0.8.9.1",
    "rerun-sdk==0.16.1",  # Pinned to work with numpy<2
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pathlib>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.5",
    "black>=21.12b0",
    "isort>=5.10.1",
    "flake8>=4.0.1",
    "mypy>=0.931",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.6.1",
    "pytest-xdist>=2.3.0",
    "pytest-timeout>=2.0.0",
    "pytest-benchmark>=3.4.1",
    "pytest-sugar>=0.9.4",
    "pre-commit>=3.3.3",
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinxcontrib-napoleon>=0.7",
]

[project.scripts]
nod-detector = "nod_detector.cli:app"

[project.urls]
homepage = "https://github.com/kkaytekin/nod_detector"
"bug-tracker" = "https://github.com/kkaytekin/nod_detector/issues"
documentation = "https://nod-detector.readthedocs.io"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["nod_detector"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = false
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.12"
